requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
PyYAML
numpy
faiss-cpu
sentence-transformers
pytest

